+++ buildings/build_floor.py
@@ -2,7 +2,7 @@ from gdpc import Block, Editor
 from gdpc.vector_tools import ivec3
 
 from buildings.building_plan import BuildingPlan
-from structures.grid import Grid
+from core.structures.grid import Grid
 
 
 def build_floor(plan: BuildingPlan, editor: Editor, build_ceiling=True):
