+++ buildings/roofs/build_roof.py
@@ -1,12 +1,12 @@
 from gdpc import Editor
 from buildings.building_plan import BuildingPlan
-from structures.grid import Grid
+from core.structures.grid import Grid
 from gdpc.vector_tools import ivec3
 from buildings.roofs.roof_component import RoofComponent, CORNER, INNER, SIDE
-from structures.legacy_directions import north, south, east, west, up
+from core.structures.legacy_directions import north, south, east, west, up
 from core.utils.vectors import y_ivec3
-from structures.nbt.build_nbt import build_nbt
-from structures.transformation import Transformation
+from core.structures.nbt.build_nbt import build_nbt
+from core.structures.transformation import Transformation
 from core.noise.rng import RNG
 
 def build_roof(plan : BuildingPlan, editor : Editor, roofComponents : list[RoofComponent], seed : int):
