+++ buildings/roofs/roof_component.py
@@ -1,6 +1,6 @@
-from structures.nbt.nbt_asset import NBTAsset
+from core.structures.nbt.nbt_asset import NBTAsset
 from core.assets.asset import asset_defaults
-from structures.legacy_directions import x_minus
+from core.structures.legacy_directions import x_minus
 
 # Roof Component Types
 SIDE = "side"
