+++ buildings/rooms/room.py
@@ -1,4 +1,4 @@
-from structures.nbt.nbt_asset import NBTAsset
+from core.structures.nbt.nbt_asset import NBTAsset
 
 class Room(NBTAsset):
     pass
