+++ core/structures/asset_structure.py
@@ -1,4 +1,4 @@
-from structures.nbt.nbt_asset import NBTAsset
+from core.structures.nbt.nbt_asset import NBTAsset
 from gdpc.vector_tools import ivec3
 
 class AssetStructure(NBTAsset):
