+++ core/structures/market/goods.py
