+++ core/structures/nbt/build_nbt.py
@@ -1,24 +1,24 @@
-from structures.nbt.convert_nbt import convert_nbt
-from structures.nbt.nbt_asset import NBTAsset
-from structures.structure import Structure
-from structures.transformation import Transformation
 from gdpc.editor import Editor
-from gdpc.block import Block
+from gdpc.vector_tools import ivec3
+
+from core.structures.nbt.convert_nbt import convert_nbt
+from core.structures.nbt.nbt_asset import NBTAsset
+from core.structures.transformation import Transformation
 from palette.palette import Palette
 from palette.palette_swap import palette_swap
-from gdpc.vector_tools import ivec3
+
 
 # Constructs an NBTAsset given an editor and transformation
 def build_nbt(
-        editor : Editor, 
-        asset : NBTAsset,
-        palette : Palette = None,
-        transformation : Transformation = None,
-        place_air : bool = False,
-        allow_non_solid_replacement : bool = False,
-    ):
+        editor: Editor,
+        asset: NBTAsset,
+        palette: Palette = None,
+        transformation: Transformation = None,
+        place_air: bool = False,
+        allow_non_solid_replacement: bool = False,
+):
     structure = convert_nbt(asset.filepath)
-    transformation = transformation or Transformation() # construct default value
+    transformation = transformation or Transformation()  # construct default value
 
     transformed_palette = transformation.apply_to_palette(structure.palette)
 
@@ -35,7 +35,7 @@ def build_nbt(
 
         # don't swap if either are null
         if asset.palette and palette:
-            block = block.copy() # I do this to avoid doubly swapping palettes
+            block = block.copy()  # I do this to avoid doubly swapping palettes
             block.name = palette_swap(block.name, asset.palette, palette)
 
         x, y, z = transformation.apply_to_point(
@@ -45,7 +45,8 @@ def build_nbt(
         )
 
         # Doesn't allow non-solid blocks to replace blocks
-        if (not allow_non_solid_replacement) and any(blocktype in block.name for blocktype in ('stairs', 'slab', 'walls', 'fence')):
+        if (not allow_non_solid_replacement) and any(
+                blocktype in block.name for blocktype in ('stairs', 'slab', 'walls', 'fence')):
             curr_block = editor.getBlock(position=(x, y, z))
 
             if 'air' not in curr_block.id:
@@ -53,13 +54,13 @@ def build_nbt(
 
         if block.name == 'minecraft:barrier':
             block.name = 'minecraft:air'
-        
-        editor.placeBlock(position=(x, y, z), block=block.to_gdpc_block(nbt)) 
+
+        editor.placeBlock(position=(x, y, z), block=block.to_gdpc_block(nbt))
 
     for (pos, entity) in structure.entities.items():
         id = entity[0]
         nbt = entity[1]
-        
+
         x, y, z = transformation.apply_to_point(
             point=pos,
             structure=structure,
@@ -67,4 +68,3 @@ def build_nbt(
         )
         summon_entity_command = f'summon {id} {x} {y} {z} {nbt}'
         editor.runCommand(summon_entity_command, position=ivec3(x, y, z))
-
