+++ core/structures/nbt/nbt_asset.py
