+++ core/structures/structure.py
@@ -1,4 +1,4 @@
-from structures.block import Block
+from core.structures.block import Block
 from gdpc.vector_tools import ivec3
 
 # A class derived from an NBT file, assets describing something built in minecraft
