+++ core/structures/transformation.py
@@ -1,7 +1,7 @@
-from structures.block import Block
-from structures.legacy_directions import x_plus, x_minus, y_plus, y_minus, z_plus, z_minus, from_text, to_text, directions
-from structures.structure import Structure
-from structures.nbt.nbt_asset import NBTAsset
+from core.structures.block import Block
+from core.structures.legacy_directions import x_plus, x_minus, y_plus, y_minus, z_plus, z_minus, from_text, to_text, directions
+from core.structures.structure import Structure
+from core.structures.nbt.nbt_asset import NBTAsset
 from gdpc.vector_tools import ivec3
 
 #region Transformation dictionaries
