+++ core/utils/sets/find_outer_points.py
@@ -1,6 +1,5 @@
-from districts.district import District
-from gdpc.vector_tools import ivec2, ivec3
-from structures.legacy_directions import cardinal, get_ivec2
+from gdpc.vector_tools import ivec2
+from core.structures.legacy_directions import cardinal, get_ivec2
 
 def find_edges(points : set[ivec2]) -> set[ivec2]:
     return {
