+++ districts/adjacency.py
@@ -1,7 +1,8 @@
 from gdpc import WorldSlice
 from districts.district import District
-from structures.legacy_directions import cardinal, get_ivec2, vector
-from gdpc.vector_tools import Rect, ivec2, distance, ivec3, Box
+from core.structures.legacy_directions import cardinal, vector
+from gdpc.vector_tools import Rect, ivec2, ivec3
+
 
 # Tells the districts what neighbours they have and why
 def establish_adjacency(world_slice : WorldSlice, district_map : list[list[District]]):
