+++ districts/district_painter.py
@@ -3,7 +3,7 @@ from gdpc.vector_tools import ivec2, ivec3
 from core.noise.rng import RNG
 from core.noise.random import choose_weighted, shuffle
 from terrain.tree import generate_tree
-from structures.legacy_directions import cardinal, get_ivec2, to_text
+from core.structures.legacy_directions import cardinal, get_ivec2, to_text
 from terrain.forest import Forest
 import time
