+++ districts/gate.py
@@ -1,11 +1,11 @@
 from gdpc import Editor, Block
 from gdpc.vector_tools import ivec2, ivec3
 from gdpc import WorldSlice
-from structures.legacy_directions import north, east, south, vector, ivec3_to_dir
+from core.structures.legacy_directions import north, east, south, vector, ivec3_to_dir
 from core.utils.geometry import is_straight_not_diagonal_ivec2
-from structures.nbt.build_nbt import build_nbt
-from structures.nbt.nbt_asset import NBTAsset
-from structures.transformation import Transformation
+from core.structures.nbt.build_nbt import build_nbt
+from core.structures.nbt.nbt_asset import NBTAsset
+from core.structures.transformation import Transformation
 from palette.palette import Palette
