+++ districts/generate_districts.py
@@ -1,7 +1,7 @@
 from core.noise.rng import RNG
 from districts.district import District
 from gdpc.vector_tools import Rect, ivec2, distance, ivec3
-from structures.legacy_directions import cardinal, vector
+from core.structures.legacy_directions import cardinal, vector
 from gdpc import WorldSlice
 from districts.adjacency import establish_adjacency
 from districts.merging_districts import merge_down
