+++ paths/build_highway.py
@@ -1,6 +1,6 @@
 from gdpc import Editor, WorldSlice, Block
 from gdpc.vector_tools import ivec3, ivec2
-from structures.legacy_directions import cardinal, get_ivec2, to_text
+from core.structures.legacy_directions import cardinal, get_ivec2, to_text
 from core.utils.bounds import is_in_bounds2d
 from core.maps.map import Map
