+++ paths/route_highway.py
@@ -1,6 +1,6 @@
 from gdpc.vector_tools import ivec3, distance
 from gdpc import Editor
-from structures.legacy_directions import all_8, vector
+from core.structures.legacy_directions import all_8, vector
 from paths.a_star import a_star, COUNTER_LIMIT_EXCEEDED
 from paths.a_star_debug import a_star_debug
 from core.utils.bounds import is_in_bounds
